readyup|refactor: Extract human output mode into its own module
Moves `runMultiKitHumanMode` out of `runCommand.ts` into a `runHumanMode.ts` of its own, renamed `runHumanMode`, and takes the five rendering helpers with it: `buildKitSegments`, `createBlockWriter`, `describeKitProvenance`, `resolveFixLocation`, and `toSummaryRow`. `runSingleKitHumanMode` becomes `runKit`, the mode being carried by the module it now sits in.

`runCommand.ts` is left with `RunCommandOptions` and the json/human dispatch. `runCommand`'s signature and `route.ts` are unchanged.
